Refactor NavigationService
