+++ src/main/java/org/chainoptim/desktop/AppModule.java
@@ -4,6 +4,7 @@ import com.google.inject.AbstractModule;
 import com.google.inject.Singleton;
 import org.chainoptim.desktop.core.main.service.CurrentSelectionService;
 import org.chainoptim.desktop.core.main.service.NavigationService;
+import org.chainoptim.desktop.core.main.service.NavigationServiceImpl;
 import org.chainoptim.desktop.core.user.repository.UserRepository;
 import org.chainoptim.desktop.core.user.repository.UserRepositoryImpl;
 import org.chainoptim.desktop.core.user.service.AuthenticationService;
@@ -17,6 +18,8 @@ import org.chainoptim.desktop.features.supplier.repository.SupplierRepositoryImp
 import org.chainoptim.desktop.features.warehouse.repository.WarehouseRepository;
 import org.chainoptim.desktop.features.warehouse.repository.WarehouseRepositoryImpl;
 import org.chainoptim.desktop.shared.fallback.FallbackManager;
+import org.chainoptim.desktop.shared.util.resourceloader.FXMLLoaderService;
+import org.chainoptim.desktop.shared.util.resourceloader.FXMLLoaderServiceImpl;
 
 import java.net.http.HttpClient;
 
@@ -28,7 +31,7 @@ public class AppModule extends AbstractModule {
     @Override
     protected void configure() {
         // Singletons
-        bind(NavigationService.class).asEagerSingleton();
+        bind(NavigationServiceImpl.class).asEagerSingleton();
         bind(FallbackManager.class).in(Singleton.class);
         bind(CurrentSelectionService.class).in(Singleton.class);
 
@@ -36,6 +39,11 @@ public class AppModule extends AbstractModule {
         bind(HttpClient.class).toInstance(HttpClient.newHttpClient());
 
         // Bind interfaces to implementations
+        // Services
+        bind(NavigationService.class).to(NavigationServiceImpl.class);
+        bind(FXMLLoaderService.class).to(FXMLLoaderServiceImpl.class);
+
+        // Repositories
         bind(AuthenticationService.class).to(AuthenticationServiceImpl.class);
         bind(UserRepository.class).to(UserRepositoryImpl.class);
         bind(ProductRepository.class).to(ProductRepositoryImpl.class);
