+++ src/main/java/org/chainoptim/desktop/core/main/service/SceneManager.java
@@ -20,16 +20,10 @@ public class SceneManager {
     @Setter
     private static Stage primaryStage;
 
-    private static Injector injector;
-
-    public static void setInjector(Injector injector) {
-        SceneManager.injector = injector;
-    }
-
     public static void loadLoginScene() throws IOException {
         FXMLLoader loader = new FXMLLoader(MainApplication.class.getResource("/org/chainoptim/desktop/core/user/AuthView.fxml"));
         // Set the controller factory to use Guice for DI
-        loader.setControllerFactory(injector::getInstance);
+        loader.setControllerFactory(MainApplication.injector::getInstance);
         Parent root = loader.load();
         Scene scene = new Scene(root, 400, 400);
         applyCss(scene, "/css/login.css");
@@ -41,7 +35,7 @@ public class SceneManager {
     public static void loadMainScene() throws IOException {
         FXMLLoader loader = new FXMLLoader(MainApplication.class.getResource("/org/chainoptim/desktop/core/main/AppView.fxml"));
         // Set the controller factory to use Guice for DI
-        loader.setControllerFactory(injector::getInstance);
+        loader.setControllerFactory(MainApplication.injector::getInstance);
         Parent root = loader.load();
         Scene mainScene = new Scene(root, 800, 600);
         applyCss(mainScene, "/css/globals.css");
