--- src/main/java/org/chainoptim/desktop/shared/util/model/NavigationItem.java
@@ -1,8 +1,0 @@
-package org.chainoptim.desktop.shared.util.model;
-
-import lombok.Data;
-
-@Data
-public class NavigationItem {
-    private String name;
-}
